Perform re-index on  set_spammer "false"
Issue #3348
The problem was that the controller performed an 'update_all' to
resolve the spammer flag(s) which doesn't trigger callbacks.
Since the flagged_with method on the user model gets triggered by a flag
change callback, the reindexing that occurs there wasn't happening.
Changing to use 'update' instead triggers a callback to reindex.